more rtow

+++ pkg/Ubuntu_16_04/rtow/install.sh
@@ -2,12 +2,12 @@ for lang in c c++ rust go scala cuda opencl crystal nim javascript haskell java 
 do
     mkdir $lang
     cd $lang
-    googler "raytracer in one weekend $lang site: github.com" --json | grep '"url":' | cut -d'"' -f4 > urls
-    cat urls | head -n1 | while read url
+    #googler "raytracer in one weekend $lang site: github.com" --json | grep '"url":' | cut -d'"' -f4 > urls
+    cat urls | head -n99 | while read url
     do
         zipurl="$url/archive/refs/heads/master.zip"
         name="$(echo $url | cut -d/ -f4).$(echo $url | cut -d/ -f5).zip"
-        wget $zipurl -O $name
+        [ -e $name ] || wget $zipurl -O $name
     done
     cd ..
 done
